+++ feature/players/shots/src/main/java/com/nicholas/rutherford/track/your/shot/feature/players/shots/selectshot/SelectShotNavigation.kt
@@ -1,4 +1,4 @@
-package com.nicholas.rutherford.track.your.shot.feature.players.shots
+package com.nicholas.rutherford.track.your.shot.feature.players.shots.selectshot
 
 interface SelectShotNavigation {
     fun popFromCreatePlayer()
