+++ feature/players/shots/src/main/java/com/nicholas/rutherford/track/your/shot/feature/players/shots/selectshot/SelectShotParams.kt
@@ -1,4 +1,4 @@
-package com.nicholas.rutherford.track.your.shot.feature.players.shots
+package com.nicholas.rutherford.track.your.shot.feature.players.shots.selectshot
 
 import com.nicholas.rutherford.track.your.shot.data.room.response.DeclaredShot
