+++ feature/players/shots/src/main/java/com/nicholas/rutherford/track/your/shot/feature/players/shots/shotscreen/ShotScreen.kt
@@ -1,0 +1,4 @@
+package com.nicholas.rutherford.track.your.shot.feature.players.shots.shotscreen
+
+class ShotScreen {
+}
