+++ feature/players/shots/src/test/java/com/nicholas/rutherford/track/your/shot/feature/players/shots/shotscreen/ShotScreenViewModelTest.kt
@@ -1,0 +1,4 @@
+package com.nicholas.rutherford.track.your.shot.feature.players.shots.shotscreen
+
+class ShotScreenViewModelTest {
+}
